Add files via upload
Hello Guys!
New here to GitHub and I have no clue in coding.
recently I started developing a trading bot with ChatGBP, after a lot of frustration I came to this point where I have a bot that is 36.57% profitable and I know I can increase it but I've came to a dead end. 
I don't know if its a play with numbers, although I've put a code that can change the look back numbers (hope it works) or I need to upgrade it even more.
ill be gladly for all of your help!

Best regards.